Change PDF output type to book
This places footnotes at the end of the chapter.

I was having problems lately (must be an asciidoctor-pdf update, but I am not sure) making the title page appear when using the default output type (article). This didn't used to be a problem for some reason, but appears to be one now. When combined with the footnote issue that was determined earlier this makes it beneficial to make the default output for the PDF to type book.